+++ packages/terra-framework-docs/CHANGELOG.md
@@ -4,12 +4,15 @@
 
 * Added
   * Added example and tests for `terra-worklist-data-grid` to select rows via mouse and keyboard.
+  * Added test to `terra-worklist-data-grid` for sticky column header feature.
 
 * Changed
   * Updated focusable cell example for `terra-worklist-data-grid` to allow text area to be editable.
   * Added examples and tests for Add & Close functionality of tabs.
   * Updated examples and tests for `terra-worklist-data-grid` to account for cell selection state being controlled by consumers.
   * Updated programmatic activation test and examples for `terra-tabs`.
+  * Updated pinned column test for `terra-worklist-data-grid` to test the divider when the table height changes.
+  * Updated examples and tests for `terra-worklist-data-grid` to handle cell selection for pinned columns.
 
 ## 1.32.0 - (August 10, 2023)