+++ packages/terra-worklist-data-grid/CHANGELOG.md
@@ -3,8 +3,10 @@
 ## Unreleased
 
 * Added
+  * Added box shadow for pinned columns in fusion theme.
   * Added ability to use spacebar on masked and non-selectable cells to select rows.
   * Added ability to treat the `terra-worklist-data-grid` component as a single tabstop.
+  * Added a sticky column header feature to `terra-worklist-data-grid`
 
 * Changed
   * Removed roving tabIndex and resolved focus management issues within cells.
@@ -12,6 +14,9 @@
 * Fixed
   * Reduced number of unnecessary rerenders in the `terra-worklist-data-grid` component.
   * Fixed cells not honoring the value of props passed by consumers.
+  * Fixed pinned column separator not adjusting size when the table height changes.
+  * Fixed pinned column cells not being selectable.
+  * Fixed pinned column border issues during scrolling.
 
 ## 0.4.0 - (August 10, 2023)